feat: http (#7)
* feat(http): introduce header types and helpers

Introduces the `HttpHeaderMapping` which represents a mapping of
header keys and values in their serialised format. Also, introduced a
few helper functions for working with header keys and mappings.

* feat(http): introduce method types and helpers

Introduces the `HttpMethod` type which is a faux-enum with methods
that are commonly used. Also, a few helper functions for working with
the normalisation of HTTP methods to ensure they are compliant.

* feat(http): introduce transport types and helpers

Introduces the `HttpTransport` type which represents the minimal
transport layer used for common responses.